requirements.txt
llama-index>=0.9.0
qdrant-client>=1.6.0
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0